classify architectures
